requirements.txt
discord.py
Flask
requests
google-generativeai
